requirements.txt
# Core project dependencies
langchain>=0.2.0
langgraph>=0.0.44
python-dotenv
requests
pydantic>=2.0

# For DAG visualization with Graphviz
graphviz>=0.20.1

# Optional: For cleaner logs
rich

# Ensure compatibility with your Python version
typing-extensions>=4.7.0; python_version < "3.11"
